backend/requirements.txt
Flask
flask-cors
python-dotenv
psycopg2-binary
together